Refactor grammar.rb
Separate Rule class from lrama/grammar.rb
